django-api/requirements.txt
asgiref==3.7.2
azure-monitor-opentelemetry==1.0.0b14
backoff==2.2.1
Deprecated==1.2.14
Django==4.2.2
djangorestframework==3.14.0
django-environ==0.10.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
importlib-metadata==6.0.1
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-django==0.39b0
opentelemetry-instrumentation-psycopg2==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-system-metrics==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
protobuf==4.23.3
psycopg2==2.9.6
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.0
wrapt==1.15.0
zipp==3.15.0
